fixed a bug where tile fetch for high zoom levels would return 404 errors from the server

+++ fluxtream-core-webapp/src/main/java/com/fluxtream/api/BodyTrackController.java
@@ -117,7 +117,7 @@ public class BodyTrackController {
     @Path("/tiles/{UID}/{DeviceNickname}.{ChannelName}/{Level}.{Offset}.json")
     @Produces({MediaType.APPLICATION_JSON})
     public String fetchTile(@PathParam("UID") Long uid, @PathParam("DeviceNickname") String deviceNickname,
-                                   @PathParam("ChannelName") String channelName, @PathParam("Level") int level, @PathParam("Offset") int offset){
+                                   @PathParam("ChannelName") String channelName, @PathParam("Level") int level, @PathParam("Offset") long offset){
         try{
             if (!checkForPermissionAccess(uid)){
                 uid = null;
